requirements.txt
asgiref==3.7.2
coverage==7.3.4
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
